\name{EACI-package}
\alias{EACI-package}
\alias{EACI}
\docType{package}
\title{
EACI - enrichment analysis taking continious inputs
}
\description{
EACI 
}
\details{
	\tabular{ll}{
		Package: \tab EACI\cr
		Type: \tab Package\cr
		License: \tab Artistic-2.0\cr
	}
}
\author{
	Ning Leng

	Maintainer: Ning Leng <lengning1@gmail.com>
}
\references{
}
\keyword{ package }
