% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eacifun.R
\name{eacifun}
\alias{eacifun}
\title{EACI model for gene set enrichment analysis}
\usage{
eacifun(X, GeneV, SetV, meanIn=NULL, sdIn=NULL, tauIn=NULL, alpha0=4, beta0=1/4,iter=2,
pred.dist=T, meanbg=NULL)
}
\arguments{
\item{X}{numeric vector for gene scores.}

\item{GeneV}{gene names. The order of GeneV should match X.}

\item{SetV}{gene set names. The order of SetV should match X. If one gene belongs to
two sets, X and GeneV should have two duplicated duplicated entries for this gene.
In SetV, these two duplicated entries should have different set names.}

\item{meanIn,sdIn,tauIn}{hyper parameters mean, sd, tau. Default is NULL. If NULL, they will be
estimated using empirical data}

\item{alpha0,beta0}{hyper parameter for prior}

\item{iter}{number of iterations}

\item{pred.dist}{posterior predictive distribution? If not, prior predictive distribution will be used.}

\item{meanbg}{background mean. If it is NULL, it will be estimated using empirical data.}
}
\value{
estimated mean sd and tau
}
\description{
EACI model for gene set enrichment analysis
}
\examples{
scores <- rnorm(20)
}
\author{
Ning Leng
}

