% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easefun.R
\name{easefun}
\alias{easefun}
\title{Calculate the DAVID EASE score for a single set}
\usage{
easefun(FAll, FMk, FIdenIn)
}
\arguments{
\item{FAll}{all genes that are considered in an analysis}

\item{FMk}{marker list of interest}

\item{FIdenIn}{genes that are identified as significant; it should be a subset of FAll.}
}
\value{
In FMk, only genes in FAll are included in the calculation
}
\description{
Calculate the DAVID EASE score for a single set
}
\examples{
f1 <- paste0("g",1:100)
f2 <- paste0("g",1:50)
f3 <- paste0("g",30:55)
res <- easefun(f1,f2,f3)
}
\author{
Ning Leng
}

