% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{fn_loadPlatformLibraries}
\alias{fn_loadPlatformLibraries}
\title{internal function - load platform library}
\usage{
fn_loadPlatformLibraries(Libraries, library.loc = library.loc)
}
\description{
internal function - load platform library
}

